pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "wechat-link"
version = "0.1.0"
description = "Unofficial Python SDK for the iLink-compatible Weixin bot protocol."
readme = { file = "README.pypi.md", content-type = "text/markdown" }
requires-python = ">=3.11"
license = { text = "MIT" }
license-files = ["LICENSE"]
authors = [{ name = "syusama" }]
keywords = ["wechat", "weixin", "ilink", "bot", "sdk", "relay", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet",
]
dependencies = ["httpx>=0.28,<1.0", "cryptography>=46.0,<47.0", "pillow>=11.3,<12.0"]

[project.urls]
Homepage = "https://github.com/syusama/wechat-link"
Repository = "https://github.com/syusama/wechat-link"
Issues = "https://github.com/syusama/wechat-link/issues"
Documentation = "https://github.com/syusama/wechat-link#readme"

[project.optional-dependencies]
dev = [
    "build>=1.2,<2.0",
    "pytest>=8.3,<9.0",
    "twine>=6.1,<7.0",
    "fastapi>=0.128,<1.0",
    "python-multipart>=0.0.20,<1.0",
    "uvicorn>=0.35,<1.0",
]
relay = [
    "fastapi>=0.128,<1.0",
    "python-multipart>=0.0.20,<1.0",
    "uvicorn>=0.35,<1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/wechat_link"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
